move status repos into statuses mod
